Basic PokemonTrainer added
I added a basic version of the PokemonTrainer named Robby. I also made a place called "pokeball" which acts very similarly to pond right now. Eventually, I want the pokeball to make the player spawn in a random part of the map (this mimics the pokemon trainer throwing the pokeball accross campus and then the character coming out of the pokeball).